feat(napi/transform): allow setting `sourceType` to `transform` (#4113)
closes #4071
